requirements.txt
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
checkdmarc==4.2.4
colour==0.1.5
dj-database-url==0.5.0
Django==3.0.3
django-appconf==1.0.3
django-autocomplete-light==3.5.1
django-bootstrap4==1.1.1
django-filter==2.2.0
django-heroku==0.3.1
django-money==1.0
django-select2==7.2.2
django-smart-selects==1.5.4
django-tables2==2.2.1
django-utils==0.0.2
dnspython==1.16.0
et-xmlfile==1.0.1
expiringdict==1.2.0
form==0.0.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
jdcal==1.4.1
Jinja2==2.11.1
json2html==1.3.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.1
openpyxl==3.0.3
pandas==1.0.0
Pillow==7.0.0
portalocker==1.5.2
psycopg2==2.8.4
publicsuffix2==2.20191221
py-moneyed==0.8.0
pyleri==1.3.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pywin32==227
requests==2.22.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.15
sqlparse==0.3.0
StyleFrame==2.0.5
timeout-decorator==0.4.1
typing==3.7.4.1
urllib3==1.25.8
whitenoise==5.0.1
xlrd==1.2.0
zipp==2.1.0
